src/test/java/com/myStore/testCases/IndexPageTest.java
/**
 * 
 */
package com.myStore.testCases;

import org.testng.Assert;
import org.testng.annotations.AfterMethod;
import org.testng.annotations.BeforeMethod;
import org.testng.annotations.Parameters;
import org.testng.annotations.Test;

import com.myStore.base.BaseClass;
import com.myStore.pageObjects.IndexPage;
import com.myStore.utility.Log;

/**
 * @author admin
 *
 */

public class IndexPageTest extends BaseClass {
	IndexPage indexPage;

	@Parameters("browser")
	@BeforeMethod(groups= {"Smoke","Sanity","Regression"})
	public void setup(String browser) {
		launchApp(browser);
	}

	@Test(groups="Smoke")
	public void verifyLogo() {
		Log.startTestCase("IndexTest Start");
		indexPage = new IndexPage();
		Log.info("Url is open");
		Assert.assertTrue(indexPage.validateLogo(), " Logo is not displayed on Indexpage");
	}

	@Test(groups="Smoke")
	public void verifyTitle() {
		Log.startTestCase("verifyTitle");
		String actTitle = indexPage.getMyStoreTitle();
		Assert.assertEquals(actTitle, "My Store");
		Log.endTestCase("verifyTitle");
	}

	@AfterMethod(groups= {"Smoke","Sanity","Regression"})
	public void tearDown() {
		getDriver().quit();
	}

}
